Refactor parseAppConfig
