+++ packages/core/src/generate/config/parseAppConfig/index.ts
@@ -1,0 +1,2 @@
+export * from "./parseAppConfig";
+export * from "./types";
